+++ bootrom/gen_rom.py
@@ -4,7 +4,8 @@ from string import Template
 import argparse
 import os.path
 import sys
-from bitstring import ConstBitStream, BitArray, BitStream
+import binascii
+
 
 parser = argparse.ArgumentParser(description='Convert binary file to verilog rom')
 parser.add_argument('filename', metavar='filename', nargs=1,
@@ -77,13 +78,10 @@ $content
 """
 
 def read_bin():
-    s = ConstBitStream(filename=filename + ".img")
-    rom = []
-    try:
-        while True:
-            rom.append(s.read("hex:8"))
-    except Exception as e:
-        pass
+
+    with open(filename + ".img", 'rb') as f:
+        rom = binascii.hexlify(f.read())
+        rom = map(''.join, zip(rom[::2], rom[1::2]))
 
 
     # align to 64 bit