+++ src/csr_regfile.sv
@@ -967,8 +967,13 @@ module csr_regfile #(
     assign tw_o             = mstatus_q.tw;
     assign tsr_o            = mstatus_q.tsr;
     assign halt_csr_o       = wfi_q;
+`ifdef PITON_ARIANE
+    assign icache_en_o      = icache_q[0];
+`else
     assign icache_en_o      = icache_q[0] & (~debug_mode_q);
+    `endif
     assign dcache_en_o      = dcache_q[0];
+
     // determine if mprv needs to be considered if in debug mode
     assign mprv             = (debug_mode_q && !dcsr_q.mprven) ? 1'b0 : mstatus_q.mprv;
     assign debug_mode_o     = debug_mode_q;