+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Official support for floating point unit
+- Added AXI-64bit adapter for write-through cache system
+- Added AXI atomic ops and exclusive access support to write-through cache system
 
 ### Changed
 
@@ -16,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Increase available RAM size on Genesys II board to 1 GiB
 - Fixed problem which decoded compressed hints as illegal instructions
 - Reduce clock frequency of FPGA to 30 MHz to accomodate FPU
+- Bugfixes in write-through cache system
 
 ### 4.0.0